+++ source/main.cpp
@@ -40,8 +40,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "SpriteSet.h"
 #include "SpriteShader.h"
 #include "TaskQueue.h"
-#include "Test.h"
-#include "TestContext.h"
+#include "test/Test.h"
+#include "test/TestContext.h"
 #include "UI.h"
 
 #include <chrono>
