+++ source/test/TestData.cpp
@@ -15,13 +15,13 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "TestData.h"
 
-#include "DataFile.h"
-#include "DataNode.h"
-#include "DataWriter.h"
-#include "Files.h"
-#include "GameData.h"
-#include "Mission.h"
-#include "UniverseObjects.h"
+#include "../DataFile.h"
+#include "../DataNode.h"
+#include "../DataWriter.h"
+#include "../Files.h"
+#include "../GameData.h"
+#include "../Mission.h"
+#include "../UniverseObjects.h"
 
 using namespace std;
